fix tracing being imported when disabled
