+++ packages/syft/src/syft/util/telemetry.py
@@ -1,10 +1,11 @@
 # stdlib
+from collections.abc import Callable
 import logging
 import os
 from typing import Any
+from typing import TypeVar
 
 # relative
-from . import trace_decorator
 from .. import __version__
 from .util import str_to_bool
 
@@ -18,10 +19,22 @@ __all__ = [
 TRACING_ENABLED = str_to_bool(os.environ.get("TRACING", "False"))
 logger = logging.getLogger(__name__)
 
+T = TypeVar("T", bound=Callable | type)
+
+
+def no_instrument(__func_or_class: T | None = None, /, *args: Any, **kwargs: Any) -> T:
+    def noop_wrapper(__func_or_class: T) -> T:
+        return __func_or_class
+
+    if __func_or_class is None:
+        return noop_wrapper  # type: ignore
+    else:
+        return __func_or_class
+
 
 def setup_instrumenter() -> Any:
     if not TRACING_ENABLED:
-        return trace_decorator.no_instrument
+        return no_instrument
 
     try:
         # third party
@@ -34,6 +47,9 @@ def setup_instrumenter() -> Any:
         from opentelemetry.sdk.resources import Resource
         from opentelemetry.sdk.trace import TracerProvider
         from opentelemetry.sdk.trace.export import BatchSpanProcessor
+
+        # relative
+        from .trace_decorator import instrument
 
         # create a resource
         resource = Resource({"syft.version": __version__})
@@ -53,10 +69,10 @@ def setup_instrumenter() -> Any:
         trace.set_tracer_provider(provider)
 
         logger.info("Added TracerProvider with BatchSpanProcessor")
-        return trace_decorator.instrument
+        return instrument
     except Exception as e:
         logger.error("Failed to import opentelemetry", exc_info=e)
-        return trace_decorator.no_instrument
+        return no_instrument
 
 
 def instrument_fastapi(app: Any) -> None:
